on finis le projet  avec auth et profil

+++ show-article.php
@@ -1,5 +1,9 @@
 <?php
+require_once __DIR__ . '/data/data-base.php';
 $articleDB = require __DIR__ . '/./data/models/ArticleDB.php';
+require_once __DIR__ . '/data/security.php';
+$currentUser = isLoggedin();
+
 
 $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 $id = $_GET['id'] ?? '';
@@ -32,10 +36,17 @@ if (!$id) {
                 <h1 class="article-title"><?= $article['title'] ?></h1>
                 <div class="separator"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
-                    <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
-                </div>
+                <?php if ($article['author']) : ?>
+                    <div class="article-author">
+                        <p><?= $article['firstname'] . ' ' . $article['lastname'] ?></p>
+                    </div>
+                <?php endif; ?>
+                <?php if ($currentUser && $currentUser['id'] === $article['author']) : ?>
+                    <div class="action">
+                        <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
+                        <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
         <?php require_once 'includes/footer.php' ?>
